requirements.txt
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
pandas==2.1.4
h5py==3.10.0
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
pyowm==3.3.0 